Move requirement reading into a requirements module
Part of #97, which asks for the requirement spec and related code to
be moved out of common.py. The RequirementSpec record, the readers for
requirements files and pyproject.toml files, the direct URL matching
they depend on, and find_required_modules now live in
pip_check_reqs/requirements.py, with their tests in
tests/test_requirements.py.
